Don't change package names for v0 files
